feat: rename iseq_loger to logger
