+++ ext/sdb/src/lib.rs
@@ -1,6 +1,5 @@
 mod gvl;
 mod helpers;
-mod iseq_logger;
 mod logger;
 mod ruby_version;
 mod stack_scanner;
